implement rotation change
